build the essential matrix
continuous system but not buid into standard formation of state space model, because it is really difficult to clarify all the non-linear term in the differential equation. Also creat several subfunction to accomplish the integral or differential function. About the integral part, two method are uesd, one is analytical method and another one is numerical method.